Audio Change
Changed the sound control system. Every Audio Source is now controlled by an Audio Mixer, instead of SoundControllScript.

NewMenu object now gets disabled in game, if the game is not paused.

There happens to be some lag spike on CanvasUpdate.PreRender when NewMenu is active - still trying to resolve the issue